Automate tables generation
- At start, tables will only be created in flask shell manually by
import the db object and call this create_all(). Now table creation is
automated.
- To support table initialization, the password of the root user in our
database has been set to empy. Strictly change the password after the
app is up and running.
- Auto matic table creation is somehow considered a bad practice.
Its only purpose for now is to speed up and completely automate the
deployment process. It can be removed in future.