Dependencies should be installed locally
- For the API module, its Python packages should be installed under the
current user home directory. This is done via --user option of pip.
- When Python site-packages are stored under home directory, one must be
cautious bind-mounting the Web's source code into the container to avoid
obscuring the installed packages. Not bind-mounting the source code
directlly into the home directory is recommended.
- For the Web module, node_modules needs to locate under the same folder
to the Web's source code. Details for safely structuring such a
environemtn is done with a trick stated in the
docker-compose.development.yml and web/Dockerfile.development.
- WORKDIR is not reliable for automatically creating non-existing
folders, because those folders are added with root permission. We should
manually create the workspace before using it with WORKDIR.